Cube location package
Previous version was set up so that the wrong file was compiling lol
This one is the correct broadcaster that is the package
The package is supposed to subscribe to tf of the cube's location, put them into a stack, and publish the average location of the cube on the map from 5 latest datapoints.
The package compiles but has not been actually tested yet.