Calibration joint for kinect
Added a floating joint(6DOF) to the tgrip.urdf.xacro file that can be used for the calibration of kinect relative to the open manipulator.
The joint kinect_adjust connects kinect to the base_footprint through link kinect_adjust_link (previously kinect as connected directly to base_footprint)
Note that gui isn't automatically enabled in rviz.